+++ src/pug/__init__.py
@@ -1,3 +1,3 @@
 """PowerPi UPS Gateway."""
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"